--- prueba.sql
@@ -1,1 +1,0 @@
-CALL BANANA_QUALITY.DEV.train_step();
