+++ src/dags/imports_train/train_func.py
@@ -1,13 +1,9 @@
-import io
 import re
-import joblib
 import pandas as pd
 
 from typing import *
 
 from snowflake.snowpark.session import Session
-from snowflake.snowpark import DataFrame
-import snowflake.snowpark.functions as F
 import snowflake.snowpark.types as T
 
 from snowflake.ml.modeling.xgboost import XGBClassifier
@@ -15,39 +11,6 @@ from snowflake.ml.modeling.metrics import roc_auc_score
 from snowflake.ml.modeling.metrics import roc_curve
 from snowflake.ml.registry import Registry
 
-## Funciones process ##
-def read_table_sf(session: Session, db_name: str, schema_name: str, table_name: str) -> DataFrame: 
-    df = session.table(f'{db_name}.{schema_name}.{table_name}')
-    return df
-
-def transform_to_numeric_target(df: DataFrame) -> DataFrame:
-    df_proc = df.withColumn(
-        "QUALITY", F.when(df["QUALITY"] == "Good", 1).otherwise(0)
-    )
-    return df_proc
-
-def write_df_to_sf(df: DataFrame, db_name: str, schema_name: str, table_name: str) -> None:
-    df.write.mode("overwrite").save_as_table(f'{db_name}.{schema_name}.{table_name}')
-
-def split_proc(session: Session, db_name: str, schema_name: str, table_name: str)-> None:
-    df_proc = read_table_sf(session, db_name, schema_name, table_name)
-    df_train, df_test = df_proc.random_split([0.8, 0.2], seed=99)
-    write_df_to_sf(df_train, db_name, schema_name, "BANANA_TRAIN")
-    write_df_to_sf(df_test, db_name, schema_name, "BANANA_TEST")
-
-
-def process_data(session: Session) -> T.Variant:
-    db = session.get_current_database().strip('"')
-    schema = session.get_current_schema().strip('"')
-
-    df_raw = read_table_sf(session, db, schema, "BANANA_QUALITY_RAW")
-   
-    df_proc = transform_to_numeric_target(df_raw)
-    write_df_to_sf(df_proc, db, schema, "BANANA_QUALITY_PROCESSED")
-
-    split_proc(session, db , schema, "BANANA_QUALITY_PROCESSED")
-    
-    return str(f'El procesamiento se realizó de manera exitosa')
 
 ## Funciones train-register ##
 def train_model(session: Session, train_table: str) -> XGBClassifier:
@@ -134,19 +97,10 @@ def train_register(session: Session) -> T.Variant:
     schema = session.get_current_schema().strip('"')
 
     xgbmodel = train_model(session, "BANANA_TRAIN")
-    xgb_file = xgbmodel.to_xgboost()
-    MODEL_FILE = 'model.joblib.gz'
-    buffer = io.BytesIO()
-    joblib.dump(xgb_file, buffer)
-    buffer.seek(0)
-
-    session.file.put_stream(buffer, f"@{schema}.ML_MODELS/{MODEL_FILE}", auto_compress=False, overwrite=True)
-
+    
     metrics_train = get_metrics(session, db, schema, "BANANA_TRAIN", xgbmodel)
     metrics_test = get_metrics(session, db, schema, "BANANA_TEST", xgbmodel)
 
-    mv = register_model(session=session, db_name=db, schema_name=schema, model=xgbmodel, model_name="BANANA_MODEL", metrics_train=metrics_train, metrics_test=metrics_test)
-
-    print("ESTE ES UN PRINT PARA VALIDAR APPENDS EN STAGE CUANDO SE EJECUTA EL PIPELINE")
+    register_model(session=session, db_name=db, schema_name=schema, model=xgbmodel, model_name="BANANA_MODEL", metrics_train=metrics_train, metrics_test=metrics_test)
 
     return str(f'El entrenamiento y registro en model registry se realizó de manera exitosa')
